Increased font size + margins.
SOCC mandates a 10pt font with 12pt leading and a text area of 6.5
inches by 9.5 inches (aka 1 inch margins all around). We were using the
ACM template which apparently defaults to 9pt font with a 7 inch by 9.5
inch text area.

This commit changes the font size and margins to obey SOCC's formatting
guidelines. The paper looks all wonky now because the sizes are all
different, but I'll fix those in later commits.

An explanation of the changes:

  - Adding 10pt to the acmart documentclass sets the font size to 10pt.
  - There wasn't an easy way that I saw to change the page size, so I
    had to edit acmart.cls directly. The class uses geometry to set the
    text area size. The inner and outer sizes (which I guess are the
    left and right margins?) were 54 pt which is 0.75 inches. I changed
    then to 72pt which is 1 inch. I also changed the columnsep from 2pc
    (whatever that is) to 1pc to shrink the space between columns to
    look more like the SOCC template.