Fixed title spacing + author abbreviation.
After increasing the font size of the paper, the title at the top left
of every page was too long. It was running into the authors at the top
right of every page. This commit adds a \\ to the middle of the title so
that it's line-wrapped at the top left of every page. I also expanded
the abbreviated author list to include everyone.