Added limitations + revised wat prov spec section.
One of the big pieces of feedback that we got from the reviewers is that
we didn't explain the limitations of wat provenance enough. This commit
adds/revises limitation subsections in the wat provenance and wat
provenance specification sections.

This commit also comments out some of the writing that the reviewers
didn't like (e.g., automatic wat provenance specification checking).