Eval formatting + revised related work.
This commit does some minor formatting to the eval section to make it
not look ugly. It also revises the related work section a bit. Mostly,
I trimmed down the related works since it was a bit long. I emphasized
the difference between network provenance and wat provenance, as a
reviewer suggested. I also added a bit to the wat provenance section
explaining that we can automatically compute why-provenance but not
wat-provenance, as a reviewer suggested.