Gutted shim overheads section.
The contribution of our paper is mostly theoretical. Wat-provenance is
an interesting formalism in and of itself. We motivate the practicality
of wat-provenance with distributed debugging, but the paper doesn't
emphasize this practical side too much. So, I think it makes sense that
we have a bit of a shorter evaluation that focuses more on
wat-provennace (i.e., succinctness and specification) as opposed to shim
overheads (which is a product of our prototype, as a reviewer pointed
out).