Expanded on difference between wat prov and GBD.
In the limitations subsection of the wat provenance section, I discussed
the difference between GDB and wat-provenance. I'm working on revising
the evaluation, and I'm going to discuss the differences between
debugging high-level bugs and lower-level bugs. I wanted this to be
clearer to readers, so I beefed up the comparison between GDB to make it
even clearer.