Don't claim debugging is easier.
One of the emphases of the camera ready is highlighting the strengths
and weaknesses of wat-provenance instead of just the strengths. As part
of this effort, I removed some strong wording from the abstract saying
that Watermelon makes debugging way easier. wat-provenance does make
debugging some kinds of bugs way WAY easier compared to other methods,
but it's a bit nuanced, so I'm avoiding the claims in the abstract.