Eval rewrite!
The goal of our initial submission was to maximize our chances of
getting accepted. The goal for our camera-ready is to make our research
as clear and as understandable as possible, even if that means
highlighting some of the weaknesses of our approach.

Our old evaluation section (as the reviewers pointed out), was bad for a
couple of reasons:

  1. It made our system seem like the best sytem of all time.

     4 orders of magnitude more concise than a state of the art system?!
     It's too good to be true!!! Not really, it is true. But, the eval
     section didn't highlight the strengths of SPADE and the weaknesses
     of wat-provenance.

  2. The numbers shows were not really important.

     Who really cares about the overheads of wat-provenance shims? The
     focus of the paper is not about wat-provenance shims, and most the
     numbers we reported were artifacts of fluent.

Our new evaluation section addresses both of these issues. It presents a
more objective comparison of wat-provenance and it measures things that
really important.

The annoying bit is that the thing that is really important is how
easy wat-provenance makes debugging, and this is not very easy at all to
measure. How do you quantify ease-of-debugging? Short of running some
user studies (which we don't have time for), there's not much good you
can do.

So, the new eval section does the best it can. It qualitatively (aka
hand wavily) compares wat-provenance to SPADE and printf debugging on a
number of metrics. wat-provenance isn't the best in all the metrics, but
it is the best for high-level debugging, which is the main problem we
set out to solve.